Feat | Netlify testing

+++ website/app/api/categories/route.ts
@@ -1,7 +1,11 @@
+import { auth } from "@/auth"
 import { NextResponse } from 'next/server'
 import { prisma } from '@/prisma'
 
-export async function GET() {
+export const GET = auth(async function GET(req) {
+    if (!req.auth) {
+        return NextResponse.json({ message: "Not authenticated" }, { status: 401 })
+    }
     try {
         const categories = await prisma.category.findMany()
         const serializedCategories = categories.map(category => ({
@@ -10,14 +14,16 @@ export async function GET() {
         }))
         return NextResponse.json(serializedCategories)
     } catch (error) {
-        console.error(error)
         return NextResponse.json({ error: 'Failed to fetch categories' }, { status: 500 })
     }
-}
+})
 
-export async function POST(request: Request) {
+export const POST = auth(async function GET(req) {
+    if (!req.auth) {
+        return NextResponse.json({ message: "Not authenticated" }, { status: 401 })
+    }
     try {
-        const { name, type } = await request.json()
+        const { name, type } = await req.json()
         const category = await prisma.category.create({
             data: {
                 categoryid: `${name.toLowerCase().replace(/\s+/g, '-')}-${Date.now()}`,
@@ -31,20 +37,21 @@ export async function POST(request: Request) {
         }
         return NextResponse.json(serializedCategory)
     } catch (error) {
-        console.error(error)
         return NextResponse.json({ error: 'Failed to create category' }, { status: 500 })
     }
-}
+})
 
-export async function DELETE(request: Request) {
+export const DELETE = auth(async function GET(req) {
+    if (!req.auth) {
+        return NextResponse.json({ message: "Not authenticated" }, { status: 401 })
+    }
     try {
-        const { categoryid } = await request.json()
+        const { categoryid } = await req.json()
         await prisma.category.delete({
             where: { categoryid },
         })
         return NextResponse.json({ success: true })
     } catch (error) {
-        console.error(error)
         return NextResponse.json({ error: 'Failed to delete category' }, { status: 500 })
     }
-}
+})
